// src/app/business/registro-cirugias/interfaces/Cirugias.ts
export default interface Cirugias{
    id:number;
    fecha: String;
    numcirugia: Number;
    numquirofano: Number;
    numExpediente: Number;
    nomPaciente: String;
    apellidoPPaciente: String;
    apellidoMPaciente: String;
    nomCirujano: String;
    nomAnestesiologo: String;
    nomResidente: String;
    nomCirugia: String;
    nomEnfeInt: String;
    nomEnfeCir: String;
    tipoAnes: String;
    inianestesia: string;
    teranestesia: string;
    medicamento: string;
    suturas: string;
    material: string;
    validCheck: string;
}